Corrected bug: possible miscalculation of COLLADA inout source index stride
Properties were renamed for clarity. indexStridePerVertex is used to
find indices for each vertex. The indexStridePerVertex is NOT
necessarily the same as the number of input sources because some
sources may share the same indices. For example, a vertex my have three
sources, (position, normal, and textCoord), but only have an index
stride of 2 because normal and texCoord always share the same index.
The stride is therefore the maximum offset of all the sources plus 1 to
account for the index at the maximum offset.